Makefile updates
Changed the makefiles to derivate the target architecture from
gcc itself, and compile all the objects into obj-$arch. This
allows several compilers to be used to build simavr, at the same
time without interferences.

This allow arm cross compilation to live in the same tree as
the x86 one...

Signed-off-by: Michel Pollet <buserror@gmail.com>