cycle_timers: Big rework
Removed the old bitfield, implemented a sorted list. Timers
get added to the list by (reverse) order of execution do acts
as a stack of timers to run and pop out.

This considerably simplifies the code, and also makes the code
that test for the next timer to run very small and efficient.

This should make cycle_timers a lot more scalable, since they are
used for many more uses than they were originaly designed for.

Signed-off-by: Michel Pollet <buserror@gmail.com>