reprap: Progress
Now the reprap board has fake stepper motors that know their position,
with endstops (they work with Marlin) also there is a way to use
arduidiot pins numbers and remap them to AVR pins..

There is a prototype "heatpot" part that will be used to collect
source and drain of heat and calculafe a temperature.

Also added mongoose embedable web server. The current plan is to
use WebGL to render the graphics of the 3D printer... altho it's
not entirely sure as yet.

Signed-off-by: Michel Pollet <buserror@gmail.com>