uart_pty: Introduced the notion of 'tap' port
The previous way to debug was 'stealing' the port by using an xterm on it..
the new way leaves the proper pty available to the application you want to use
but still allow you to have a 'tap' port that sees both the sides of the
communication om the other pty.

You can also type stuff obviously, and it does some cr/lf conversions.

/tmp/simavr-uart0 : will be the normal fake serial port
/tmp/simavr-uarttap0 : is the one to launch an xterm on

This allow you to really see the serial traffic as it happends with both
requests and replies.

Signed-off-by: Michel Pollet <buserror@gmail.com>