cycle_counters: add initial implementation
Cycle counters allow measuring the execution time of code blocks.
This is particularly convenient when used with MCU section macros.
The counters attempt to account for overhead incurred by the
instructions used to communicate with simavr. This even works when
nesting counters.
The method employed is quite limited though and not accurate in all
cases.

Another shortcoming of the current implementation is that any time spent
servicing IRQs is included in the measurement.