vcd: Extensive update; added support for VCD Input
Massive update here. Now allow reading from a VCD file, as long as the
traces are named in a way that allows us to connect an IRQ. Typically,
you can send pin values to for example PORTB5 by naming the trace
'iogB_5' -- you can use sigrock for example to capture traces from
real hardware, and 'replay' it in simavr.

Also since I was there, reworked the whole module to use a 'proper' FIFO
and simplified the output generation too.

There are a few 'TODO' left, but it's functional.

Signed-off-by: Michel Pollet <buserror@gmail.com>