parts: Build a simavr virtual parts library
An external project using simavr can link against libsimavr.
However, there is no support for using virtual parts from the
examples folder in the external project.

This commit adds support for building a parts library, which the
external project can link against. Presently this library is fairly
small, but as the number of parts grows this approach may need
to be changed. A longer term solution could be to provide an API as
part of the simavr library which can manage the parts. Simavr could
dynamically load and provide parts as required by the external
project.

Like libsimavr, this patch makes no attempt to make the version easy
to set. That can be covered under a different PR.