Timer and io independen use of avr core, parts link library, run_avr version info
The cycle timer and io methods are disentangled from the core avr
structure. This allows to add periphery modules and provide bindings for
script languages like python without the need to pass the core avr
structure to the simulated periphery. In addition it is possible that
periphery has its own subtimer which is halted as soon as the main cycle
timer is halted for example in debugger. To keep changes as minimal as
possible original methods are conerted to wrapper methods where
possible.
Further it was necessary to add avr_remove_signal wich takes the id of
the signal to be removed which was previously returned by avr_add_sig.
Changed avr_add_signal to internally call avr_add_sig and return 0
(success) for any return value >= 0

run_avr did not print the version of the sim_avr package it was built
for, that makes it difficult for dependent project build systems (eg.
cmake) to verify that the version reported by pkg-config matches the
one sim_avr was built from, and systems where pkg-config is not
installed at all checking for matching version is not possible.

Makefile now autogenerates as versioninfo.h file which contains all the
necessary defines and properly updates the pc files for pkg-info

No link library existed exporting the peripheral parts which are provided
by the simavr package. Such a library would allow dependent packages
such as simutron, uc-sim (my own) to utilize the parts without the need
to embedd the sources of sim_avr within their own package.

Moved the font.h file form borad_hd7480 directory to examples parts
directory to ensure it can be compiled and thus packages along with the
parts library. Added simavr-parts.pc file for pkg-config to ensure the
library and required compile and link flags can be found by dependent
projects. Updated test/makefile to build parts libary.