Add GitHub Workflow for Linux (attempt MacOS)
This triggers a GitHub Action that gets the dependencies for avr-gcc
(and also freeglut) in order to run the `make` on Linux.

After the make is complete, it runs the suggested test command from
the README.md.  Currently this demonstrates that the output does
not match what is stated.  This could be evolved into a test that
uses the continuous integration to enforce that the README.md output
stays in sync with the actual program output.

It includes an attempt to build on a Mac container.  It gets an
error of "fatal: No names found, cannot describe anything"...which
someone familiar with MacOS who has seen a successful build would
be more qualified to try and fix than I would be.