Partial merge of up to date avrlib headers
This is a complete mess. I have 4 copies, all of them with differences,
unclear copyright dates and content. The most recent seems ... not to be
the most recent for many files!

So this is mostly a merge of what I had, what's on Debian, and what's on
the Arduino base toolchain... Ewww

Signed-off-by: Michel Pollet <buserror@gmail.com>