flash+eeprom: fix ioctl handling
The flash ioctl handler unconditionally reported the request as
processed even when it was skipped. This lead to eeprom ioctl handler
not getting called at all if it was registered later than the flash.

The eeprom ioctl handler reported the request as not processed even if
it was, so fix that too while at it.

This bug resulted in the emulated firmware seeing blank EEPROM even when
the contents was provided (as a section in ELF or as a separate HEX
file) and r/w access via GDB not working.

Signed-off-by: Paul Fertser <fercerpav@gmail.com>