Fix two UAF errors in read_ihex_chunks()
1. Linked list uses realloc but did not update
the pointer in the previous linked list item
resulting in Use-After-Free error. A backlink_p
local variable is introduced to update the pointer
address after a reallocation.

2. The local variable chunks_p was not updated
after realloc which also resulted in UAF when
allocating a new element to the linked list.
A refreshing of the address was added after realloc.