Reset ANSI colors to default before newline (#556)
UART output is printed in color as follows:

    AVR_LOG(avr, LOG_OUTPUT, FONT_GREEN "%s\n" FONT_DEFAULT, ...);

Trace and error messages are formatted similarly, with FONT_DEFAULT
following the newline. As these messages go to stdout, which defaults to
being line-buffered, the FONT_DEFAULT ANSI escape code is not output to
the terminal right away, and stays in the stdout output buffer. If this
is followed by a message printed to stderr, which is unbuffered, that
second message will be unduly colored.

Fix this by sending FONT_DEFAULT before the newline. While we are at it,
remove a stray FONT_DEFAULT from a LOG_WARNING uncolored message.