CI: upgrade aarch64 runner to ubuntu-24.04
Ubuntu 22.04 ships a version of e2fsprogs that is too old to support the
orphan_file filesystem feature, and this prevents the build-aarch64 job to
resize the VM image used to run the build. This is causing build
failures on “No space left on device” errors.

Fix the failures by upgrading to a more recent runner image.

Fixes: https://github.com/buserror/simavr/actions/runs/18200025079/job/51816354774